More robust "are we running in a TTY" check
Previously the return value of the setfont program was used, which sets
the font used in a Linux console.  In a PTY (e.g. terminal emulator), it
generally fails, but who knows, it might not.

This commit makes the check a little closer to the source of the
information, by checking what device stdin is connected to: is it one of
/dev/tty* or something else?  If it's /dev/tty*, we're in a console and
deploy the font hack.

May help with karlstav/cava#68.