Added Hann windowing and proper ring buffer
before input frames from the audio process where just added to the
buffer front to back and then wrapped around. Now samples are added
to the front of the buffer and the older are pushed backwards. This
removes some noise that could be seen earlier.

Hann windowing is also added to detrend the audio data. This removes
a lot of noise that we had earlier.

Also adapted the audio threads to read the same number of bytes as the
treble fftw buffer, since no more is needed (and less will reduce precision).

Reading more will cause crash, this needs to be looked at for shmem!