Implement different orientations of output for SDL
Currently cava only supports bars that are aligned to the bottom of the
window and point upwards. This commit allows them to be aligned to any
edge of the window.

For now it is only implemented this for SDL, because it was the simplest
to, but I may also do it for ncurses output.