increase max limit of bars to 512 per channel (#636)
previously this was 512 max total, this is now 512 per output channel

for raw it was only 256 total due to a bar_spacing of 1 used in the calculation
this is also fixed

so if stereo output is configured there can be 1024 bars
the limit is in the audio input buffer for the FFT
for higher sample rates a higher limit is theoretically possible,
this can be looked at later if the intereset is there

sdl_glsl output is still limited to 512 due to a limit on uniforms
in openGL, there are ways to resolve this if needed (look at the android shader).
But that would break all existing shaders.