pipewire should not handle SIGINT and SIGTERM #639
those flags where maybe added in an attempt to exit the pipewire loop
more gracefully, but the signals are already handled in the main loop.
So it would be random if the pipewire loop or the main loop captured
them.

this commit removes the signal handling from the pipewire loop and
extends main loop signal handling to exit normally instead
of raising signal to self immediately, this means stoping audio thread
and free all buffers