combine mid and treble FFT input buffers
The point of having separate input buffers for bass, mids and treble
was only ever to capture lowest bass without sacrificing the
responsivness of the rest of the spectrum. In the process I also
made a separate buffer for treble, but as it turns out the buffer is way
too small to be usefull and responsiveness is just as good with the mid
buffer. So this commit removes the treble buffer and uses the mid buffer
for all frquencies over the bass cut off (100Hz).

The fft input buffer size is then 4096 samples for 44100 and 48000
samples rate. (double for 96k, half for 22k and so forth) although that
should theoretically be able to capture frequencies down towards 10Hz it
is not that good at it. It will also have a minimum 10 Hz resolution
between the bars. The response time is around 100ms (or rather
samples that are 100ms old will still be visualized). This will
introduce some more inertia to the visualization, I have tuned down the
noise filter to counter for this.

The bass buffer is twice as large, and will go twice as low in the
spectrum and be twice as slow.

some adjustments and fixes where also made to standard eq and cut off frequencies

use correct buffersizes for each band in the hardcoded eq

adjust hardcoded eq to match old behaviour, less bass

update cavacore blueprint, changed slightly due to adjustments