Reduce upper frequency cutoff to 8kHz and increase eq for highest frequencies
Not many particular distinguishable sounds above 8kHz excpet for maybe
s and f consanant sounds and some nuances in hi-hats and cymbals.
These are not very interseting and cause high peaks
in some music. Reducing the upper frequency cutoff to 8kHz also allows
for upping the hardcoded eq for the higher frequencies which gives a
better distribution.